requirements.txt

accelerate>=0.26.0
black
cv2>=4.8.0
diffusers>=0.27.0
huggingface-hub>=0.23.0
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0
tqdm>=4.66.0
scipy>=1.13.0
numpy>=1.26.0
Pillow>=10.2.0
